♻️ refactor: 모달 컴포넌트 분리(add,detail)
